tests/wpt/tests/html/canvas/offscreen/path-objects/2d.path.ellipse.basics.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.path.ellipse.basics</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.path.ellipse.basics</h1>
<p class="desc">Verify canvas throws error when drawing ellipse with negative radii.</p>


<script>
var t = async_test("Verify canvas throws error when drawing ellipse with negative radii.");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.ellipse(10, 10, 10, 5, 0, 0, 1, false);
  ctx.ellipse(10, 10, 10, 0, 0, 0, 1, false);
  ctx.ellipse(10, 10, -0, 5, 0, 0, 1, false);
  assert_throws_dom("INDEX_SIZE_ERR", function() { ctx.ellipse(10, 10, -2, 5, 0, 0, 1, false); });
  assert_throws_dom("INDEX_SIZE_ERR", function() { ctx.ellipse(10, 10, 0, -1.5, 0, 0, 1, false); });
  assert_throws_dom("INDEX_SIZE_ERR", function() { ctx.ellipse(10, 10, -2, -5, 0, 0, 1, false); });
  ctx.ellipse(80, 0, 10, 4294967277, Math.PI / -84, -Math.PI / 2147483436, false);
  t.done();

});
</script>
